Move test for Reorg from GenerateIRTest to NativeTest
